Move update_document_metadata() into models
