+++ h/api/storage.py
@@ -118,66 +118,10 @@ def create_annotation(request, data):
     # annotation.updated get created.
     request.db.flush()
 
-    update_document_metadata(
+    models.update_document_metadata(
         request.db, annotation, document_meta_dicts, document_uri_dicts)
 
     return annotation
-
-
-def update_document_metadata(session,
-                             annotation,
-                             document_meta_dicts,
-                             document_uri_dicts):
-    """
-    Create and update document metadata from the given annotation.
-
-    Document, DocumentURI and DocumentMeta objects will be created, updated
-    and deleted in the database as required by the given annotation and
-    document meta and uri dicts.
-
-    :param annotation: the annotation that the document metadata comes from
-    :type annotation: h.api.models.Annotation
-
-    :param document_meta_dicts: the document metadata dicts that were derived
-        by validation from the "document" dict that the client posted
-    :type document_meta_dicts: list of dicts
-
-    :param document_uri_dicts: the document URI dicts that were derived by
-        validation from the "document" dict that the client posted
-    :type document_uri_dicts: list of dicts
-
-    """
-    documents = models.Document.find_or_create_by_uris(
-        session,
-        annotation.target_uri,
-        [u['uri'] for u in document_uri_dicts],
-        created=annotation.created,
-        updated=annotation.updated)
-
-    if documents.count() > 1:
-        document = models.merge_documents(session,
-                                          documents,
-                                          updated=annotation.updated)
-    else:
-        document = documents.first()
-
-    document.updated = annotation.updated
-
-    for document_uri_dict in document_uri_dicts:
-        models.create_or_update_document_uri(
-            session=session,
-            document=document,
-            created=annotation.created,
-            updated=annotation.updated,
-            **document_uri_dict)
-
-    for document_meta_dict in document_meta_dicts:
-        models.create_or_update_document_meta(
-            session=session,
-            document=document,
-            created=annotation.created,
-            updated=annotation.updated,
-            **document_meta_dict)
 
 
 def update_annotation(session, annotation_id, data):
@@ -221,7 +165,7 @@ def update_annotation(session, annotation_id, data):
     if document:
         document_uri_dicts = document['document_uri_dicts']
         document_meta_dicts = document['document_meta_dicts']
-        update_document_metadata(
+        models.update_document_metadata(
             session, annotation, document_meta_dicts, document_uri_dicts)
 
     return annotation
