+++ h/api/test/storage_test.py
@@ -451,181 +451,6 @@ class TestCreateAnnotation(object):
         }
 
 
-class TestUpdateDocumentMetadata(object):
-
-    def test_it_calls_find_or_create_by_uris(self,
-                                             annotation,
-                                             models,
-                                             session):
-        document_uri_dicts = [
-            {
-                'uri': 'http://example.com/example_1',
-                'claimant': 'http://example.com/claimant',
-                'type': 'type',
-                'content_type': None,
-            },
-            {
-                'uri': 'http://example.com/example_2',
-                'claimant': 'http://example.com/claimant',
-                'type': 'type',
-                'content_type': None,
-            },
-            {
-                'uri': 'http://example.com/example_3',
-                'claimant': 'http://example.com/claimant',
-                'type': 'type',
-                'content_type': None,
-            },
-        ]
-
-        storage.update_document_metadata(session,
-                                         annotation,
-                                         [],
-                                         document_uri_dicts)
-
-        models.Document.find_or_create_by_uris.assert_called_once_with(
-            session,
-            annotation.target_uri,
-            [
-                'http://example.com/example_1',
-                'http://example.com/example_2',
-                'http://example.com/example_3',
-            ],
-            created=annotation.created,
-            updated=annotation.updated,
-        )
-
-    def test_it_calls_merge_documents(self, annotation, session, models):
-        """If it finds more than one document it calls merge_documents()."""
-        models.Document.find_or_create_by_uris.return_value = mock.Mock(
-            count=mock.Mock(return_value=3))
-
-        storage.update_document_metadata(session, annotation, [], [])
-
-        models.merge_documents.assert_called_once_with(
-            session,
-            models.Document.find_or_create_by_uris.return_value,
-            updated=annotation.updated)
-
-    def test_it_calls_first(self, annotation, session, models):
-        """If it finds only one document it calls first()."""
-        models.Document.find_or_create_by_uris.return_value = mock.Mock(
-            count=mock.Mock(return_value=1))
-
-        storage.update_document_metadata(session, annotation, [], [])
-
-        models.Document.find_or_create_by_uris.return_value\
-            .first.assert_called_once_with()
-
-    def test_it_updates_document_updated(self, annotation, session, models):
-        yesterday = "yesterday"
-        document = models.merge_documents.return_value = mock.Mock(
-            updated=yesterday)
-        models.Document.find_or_create_by_uris.return_value.first\
-            .return_value = document
-
-        storage.update_document_metadata(session, annotation, [], [])
-
-        assert document.updated == annotation.updated
-
-    def test_it_calls_create_or_update_document_uri(self,
-                                                    session,
-                                                    annotation,
-                                                    models):
-        models.Document.find_or_create_by_uris.return_value.count\
-            .return_value = 1
-
-        document_uri_dicts = [
-            {
-                'uri': 'http://example.com/example_1',
-                'claimant': 'http://example.com/claimant',
-                'type': 'type',
-                'content_type': None,
-            },
-            {
-                'uri': 'http://example.com/example_2',
-                'claimant': 'http://example.com/claimant',
-                'type': 'type',
-                'content_type': None,
-            },
-            {
-                'uri': 'http://example.com/example_3',
-                'claimant': 'http://example.com/claimant',
-                'type': 'type',
-                'content_type': None,
-            },
-        ]
-
-        storage.update_document_metadata(session,
-                                         annotation,
-                                         [],
-                                         document_uri_dicts)
-
-        assert models.create_or_update_document_uri.call_count == 3
-        for doc_uri_dict in document_uri_dicts:
-            models.create_or_update_document_uri.assert_any_call(
-                session=session,
-                document=models.Document.find_or_create_by_uris.return_value.first.return_value,
-                created=annotation.created,
-                updated=annotation.updated,
-                **doc_uri_dict
-            )
-
-    def test_it_calls_create_or_update_document_meta(self,
-                                                     annotation,
-                                                     session,
-                                                     models):
-        models.Document.find_or_create_by_uris.return_value.count\
-            .return_value = 1
-
-        document_meta_dicts = [
-            {
-                'claimant': 'http://example.com/claimant',
-                'claimant_normalized':
-                    'http://example.com/claimant_normalized',
-                'type': 'title',
-                'value': 'foo',
-            },
-            {
-                'type': 'article title',
-                'claimant_normalized':
-                    'http://example.com/claimant_normalized',
-                'value': 'bar',
-                'claimant': 'http://example.com/claimant',
-            },
-            {
-                'type': 'site title',
-                'claimant_normalized':
-                    'http://example.com/claimant_normalized',
-                'value': 'gar',
-                'claimant': 'http://example.com/claimant',
-            },
-        ]
-
-        storage.update_document_metadata(session,
-                                         annotation,
-                                         document_meta_dicts,
-                                         [])
-
-        assert models.create_or_update_document_meta.call_count == 3
-        for document_meta_dict in document_meta_dicts:
-            models.create_or_update_document_meta.assert_any_call(
-                session=session,
-                document=models.Document.find_or_create_by_uris.return_value.first.return_value,
-                created=annotation.created,
-                updated=annotation.updated,
-                **document_meta_dict
-            )
-
-    @pytest.fixture
-    def annotation(self):
-        return mock.Mock(spec=Annotation())
-
-    @pytest.fixture
-    def session(self):
-        return mock.Mock(spec=db.Session)
-
-
 @pytest.mark.usefixtures('models',
                          'update_document_metadata')
 class TestUpdateAnnotation(object):
@@ -867,4 +692,4 @@ def transform(patch):
 
 @pytest.fixture
 def update_document_metadata(patch):
-    return patch('h.api.storage.update_document_metadata')
+    return patch('h.api.storage.models.update_document_metadata')
